kun virusscan hvis det er vedlegg
